requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.5
pydantic>=2.0.0
typing-extensions>=4.5.0
python-multipart>=0.0.6
starlette>=0.27.0
